# openapi.yaml
openapi: 3.1.0
...
info:
  title: tdb-search
  version: 0.1.0
  description: >
    Standalone semantic search engine — push-indexed, versioned per commit, with
    vector, full-text, and hybrid search (hybrid is the default).
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Default local bind ([::]:8080).
...
security:
  - adminSecret: []
paths:
...
  /last-indexed:
    get:
      tags: [indexing]
      operationId: getLastIndexed
      summary: Report the last-indexed commit for a (domain, branch).
      description: >
        A client calls this to learn where the engine is up to, then computes the
        delta `last → head` to push. Returns `commit: null` if the
        (domain, branch) has never been indexed.
      parameters:
        - $ref: '#/components/parameters/Domain'
        - $ref: '#/components/parameters/Branch'
      responses:
        '200':
          description: The last-indexed commit (or null if none).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LastIndexed' }
              examples:
                indexed:   { value: { branch: main, commit: "abc123", version: 7 } }
                unindexed: { value: { branch: main, commit: null, version: 0 } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /push:
    post:
      tags: [indexing]
      operationId: pushDelta
      summary: Ingest an NDJSON operation delta for a target commit (async).
      description: >
        Body is an NDJSON operation stream (one operation per line; see
        OperationLine). Parsed and processed incrementally — never buffered whole.
        The engine chunks each rendered string, embeds it, upserts/deletes chunk
        rows, and on completion binds `target_commit` to the resulting snapshot.
        Returns a task id to poll via `/check`. `parent_commit` comes from the
        `/last-indexed` handshake.
      parameters:
        - $ref: '#/components/parameters/Domain'
        - $ref: '#/components/parameters/Branch'
        - name: target_commit
          in: query
          required: true
          description: The commit this delta produces.
          schema: { type: string }
        - name: parent_commit
          in: query
          required: false
          description: >
            The commit the delta is computed from (the prior last-indexed).
            Absent means the first index of this lineage (every doc is Inserted).
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/x-ndjson:
            # The wire body is one OperationLine per line (newline-delimited),
            # parsed incrementally. Modelled here as the item schema so the
            # contract for each line is explicit and validatable.
            schema:
              $ref: '#/components/schemas/OperationLine'
            examples:
              delta:
                value: |
                  {"op":"Inserted","id":"terminusdb:///star-wars/People/20","string":"The person's name is Yoda. ..."}
                  {"op":"Changed","id":"terminusdb:///star-wars/People/22","string":"The person's name is Boba Fett. ..."}
                  {"op":"Deleted","id":"terminusdb:///star-wars/People/21"}
                  {"op":"Error","message":"Failed to render terminusdb:///star-wars/People/99"}
      responses:
        '200':
          description: Push accepted; indexing runs asynchronously. Poll `/check`.
          content:
            text/plain:
              schema: { type: string, description: Opaque task id. }
              examples: { taskId: { value: "task-7f3a9c" } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: A push for this (domain, branch) is already in progress (concurrency guard).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /check:
    get:
      tags: [indexing]
      operationId: checkTask
      summary: Poll the status of a push/index task.
      parameters:
        - name: task_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Task status (Pending or Complete).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TaskPending'
                  - $ref: '#/components/schemas/TaskComplete'
              examples:
                pending:  { value: { status: Pending, percentage: 42.5 } }
                complete: { value: { status: Complete, indexed_documents: 1284, skipped: [] } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown task id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: >
            The task failed (systemic error). Body is the error message; the
            failure is never masked as success.
          content:
            text/plain:
              schema: { type: string }
  /assign:
    post:
      tags: [indexing]
      operationId: assignCommit
      summary: Point a target commit at a source commit's index (no recompute).
      description: >
        Binds `target_commit` to the existing snapshot of `source_commit` — used
        when a new commit changes nothing indexable. No embedding, no data
        movement. This mutates state (it creates a commit→snapshot binding), so it
        is POST. Idempotent: re-assigning the same pair is a no-op.
      parameters:
        - $ref: '#/components/parameters/Domain'
        - name: source_commit
          in: query
          required: true
          schema: { type: string }
        - name: target_commit
          in: query
          required: true
          schema: { type: string }
      responses:
        '204': { description: Assigned. }
        '400':
          description: Bad request (e.g. unknown source commit).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────────── search ────────────────────────────────────
  /search:
    get:
      tags: [search]
      operationId: searchGet
      summary: Semantic / full-text / hybrid search via query string (hybrid default).
      description: >
        Read-only, cacheable search. The query text is the `q` parameter. Use this
        for simple, link-safe, cacheable searches; use POST for a structured JSON
        request (long queries, programmatic composition). `doc_type`/`doc_id` are
        repeated params (e.g. `?doc_type=A&doc_type=B`), AND-ed across the two
        sets, OR within each. Results paginate with `start` (default 0) and
        `count` (default 50). Staleness is reported via the `TerminusDB-Data-Version`
        response header.
      parameters:
        - $ref: '#/components/parameters/Domain'
        - $ref: '#/components/parameters/Commit'
        - name: q
          in: query
          required: true
          description: The query text (required on GET).
          schema: { type: string }
          example: "Wise old man"
        - $ref: '#/components/parameters/Start'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Mode'
        - $ref: '#/components/parameters/DocType'
        - $ref: '#/components/parameters/DocId'
        - $ref: '#/components/parameters/Snippet'
        - $ref: '#/components/parameters/ExpectedDataVersion'
      responses:
        '200': { $ref: '#/components/responses/SearchOk' }
        '400': { $ref: '#/components/responses/SearchBadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NoIndexedLineage' }
        '409': { $ref: '#/components/responses/DimensionMismatch' }
        '503': { $ref: '#/components/responses/EmbeddingBackendUnavailable' }
        '500': { $ref: '#/components/responses/UpstreamError' }
    post:
      tags: [search]
      operationId: search
      summary: Semantic / full-text / hybrid search via JSON body (hybrid default).
      description: >
        Structured search. The JSON body carries the query and all parameters. In
        `hybrid` (default) the same query text is embedded for the vector side and
        used as full-text terms; results dedup to documents (best chunk per
        document).


        Parameter source & precedence: every parameter may be given EITHER as a
        query parameter OR in the JSON body. The JSON body takes precedence — if a
        field is present in the body, the same-named query parameter is discarded
        (not merged). This lets a caller set defaults in the URL and override them
        in the body. `domain`, `commit` and `q` must be present in the body OR the
        query — absent in both gives 400. Results paginate with `start` (default
        0) and `count` (default 50). Staleness is reported via the
        `TerminusDB-Data-Version` response header — the response body stays the
        bare array.


        Every query parameter below is optional at the query layer on POST,
        because the JSON body may supply it (and wins).
      parameters:
        - name: domain
          in: query
          required: false
          description: Optional on POST — body `domain` wins (see SearchRequest).
          schema: { type: string }
        - name: commit
          in: query
          required: false
          description: Optional on POST — body `commit` wins (see SearchRequest).
          schema: { type: string }
        - $ref: '#/components/parameters/Q'
        - $ref: '#/components/parameters/Start'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Mode'
        - $ref: '#/components/parameters/DocType'
        - $ref: '#/components/parameters/DocId'
        - $ref: '#/components/parameters/Snippet'
        - $ref: '#/components/parameters/ExpectedDataVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchRequest' }
            examples:
              minimal:
                value: { domain: "admin/star_wars", commit: "abc123", q: "Wise old man" }
              filtered:
                value:
                  domain: "admin/star_wars"
                  commit: "abc123"
                  q: "Wise old man"
                  mode: hybrid
                  start: 0
                  count: 20
                  doc_type: ["People"]
                  doc_id: ["terminusdb:///star-wars/People/20"]
      responses:
        '200': { $ref: '#/components/responses/SearchOk' }
        '400': { $ref: '#/components/responses/SearchBadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NoIndexedLineage' }
        '409': { $ref: '#/components/responses/DimensionMismatch' }
        '503': { $ref: '#/components/responses/EmbeddingBackendUnavailable' }
        '500': { $ref: '#/components/responses/UpstreamError' }
  /similar:
    get:
      tags: [search]
      operationId: similar
      summary: Documents similar to a known indexed document.
      description: >
        Looks up the document's vector(s) by id, then runs a vector search. Same
        response shape and staleness header as `/search`; paginates with `start`
        and `count`.
      parameters:
        - $ref: '#/components/parameters/Domain'
        - $ref: '#/components/parameters/Commit'
        - name: id
          in: query
          required: true
          description: Full document IRI to find neighbours of.
          schema: { type: string }
        - $ref: '#/components/parameters/Start'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/DocType'
        - $ref: '#/components/parameters/Snippet'
        - $ref: '#/components/parameters/ExpectedDataVersion'
      responses:
        '200':
          description: Ranked neighbours (same shape/header as /search).
          headers:
            TerminusDB-Data-Version: { $ref: '#/components/headers/DataVersion' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResults' }
        '400':
          description: The id is not present in the index, or a param is invalid.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NoIndexedLineage' }
        '500': { $ref: '#/components/responses/UpstreamError' }
  /duplicates:
    get:
      tags: [search]
      operationId: duplicates
      summary: Corpus-wide near-duplicate candidate pairs (bounded).
      description: >
        For each indexed point, find its nearest neighbour; pairs with distance
        below `threshold` are returned (lower id first). Bounded by `threshold`, a
        candidate cap, and pagination — never an unbounded all-pairs scan. If a
        safe bounded implementation is unavailable, returns 501 rather than a scan
        that does not scale.
      parameters:
        - $ref: '#/components/parameters/Domain'
        - $ref: '#/components/parameters/Commit'
        - name: threshold
          in: query
          required: false
          description: Max distance for a pair to count as a near-duplicate.
          schema: { type: number, format: float, default: 0.0 }
        - $ref: '#/components/parameters/Start'
        - $ref: '#/components/parameters/Count'
      responses:
        '200':
          description: Array of [id1, id2] pairs (lower id first).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DuplicatePairs' }
              examples:
                pairs:
                  value:
                    - ["terminusdb:///star-wars/People/10", "terminusdb:///star-wars/People/57"]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NoIndexedLineage' }
        '501':
          description: Bounded duplicates unavailable — never an unbounded scan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ───────────────────────────── ops ───────────────────────────────────────
  /statistics:
    get:
      tags: [ops]
      operationId: statistics
      summary: Engine statistics.
      description: >
        Operational counters across all indexed data. Counters are
        best-effort/advisory (may be approximate under concurrency).
      responses:
        '200':
          description: Statistics.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Statistics' }
              examples:
                stats:
                  value:
                    domains: 3
                    branches: 9
                    indexed_commits: 412
                    documents: 18044
                    chunks: 51234
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    adminSecret:
      type: http
      scheme: basic
      description: >
        Shared admin secret (HTTP Basic). Authentication only — no RBAC; the
        engine is a trusted component on a private network. Checked on every
        functional endpoint with a constant-time comparison; missing/wrong gives
        401. The only per-request credential; the embedding provider key, if any,
        is the engine's own server-side configuration.
  parameters:
    Domain:
      name: domain
      in: query
      required: true
      description: >
        The database to address, as a graphspec. The repository segment is always
        part of the address and defaults to `local`:
          • `org/db`                          → `org/db/local/branch/main`
          • `org/db/<repo>`                   → `org/db/<repo>/branch/main`
          • `org/db/<repo>/branch/<branch>`   → that branch
          • `org/db/<repo>/commit/<commit>`   → that snapshot, directly
        A branch is resolved to a concrete commit by the caller (TerminusDB owns
        branch heads); the searched snapshot is always explicit (`commit`), never
        an implicit "latest".
      schema: { type: string }
      example: admin/star_wars
    Branch:
      name: branch
      in: query
      required: true
      description: Branch name.
      schema: { type: string, default: main }
...